feat: verno-167 - building now
- implementing guice
- restructuring project

+++ server/src/main/java/ch/verno/server/service/base/AbstractEntityService.java
@@ -1,6 +1,9 @@
 package ch.verno.server.service.base;
 
+import ch.verno.contract.dto.response.base.delete.DeleteResponse;
 import ch.verno.contract.dto.table.base.BaseDto;
+import ch.verno.lib.exception.ExceptionUtil;
+import ch.verno.lib.lang.ObjectUtil;
 import ch.verno.server.mapper.base.IEntityMapper;
 import ch.verno.server.repository.base.IEntityRepository;
 import jakarta.annotation.Nonnull;
@@ -37,12 +40,25 @@ public abstract class AbstractEntityService<
     return mapper;
   }
 
+  @Override
+  @Transactional(readOnly = true)
+  public boolean existsById(@Nonnull final ID id) {
+    return repository.existsById(id);
+  }
+
   @Nonnull
   @Override
   @Transactional(readOnly = true)
   public Optional<DTO> findById(@Nonnull final ID id) {
     return repository.findById(id)
             .map(mapper::toSimpleDto);
+  }
+
+  @Nonnull
+  @Override
+  @Transactional(readOnly = true)
+  public DTO findByIdRequired(@Nonnull final ID id) {
+    return getMapper().toSimpleDto(repository.findById(id).orElseThrow(ExceptionUtil::toEntityNotFoundException));
   }
 
   @Nonnull
@@ -88,20 +104,28 @@ public abstract class AbstractEntityService<
     return mapper.toSimpleDto(entity);
   }
 
+  @Nonnull
   @Override
   @Transactional
-  public boolean deleteById(@Nonnull ID id) {
-    return repository.deleteById(id);
-  }
-
-  @Override
-  @Transactional
-  public void delete(@Nonnull final DTO dto) {
-    if (dto.getId() == null) {
-      throw new ServiceException("Cannot delete entity with null id");
+  public DeleteResponse deleteById(@Nonnull final ID id) {
+    if (ObjectUtil.isEmpty(id)) {
+      return DeleteResponse.emptyId();
     }
 
-    repository.deleteById(dto.getId());
+    final var result = repository.deleteById(id);
+    return result ? DeleteResponse.success() : DeleteResponse.faulty();
+  }
+
+  @Nonnull
+  @Override
+  @Transactional
+  public DeleteResponse delete(@Nonnull final DTO dto) {
+    if (dto.getId() == null || ObjectUtil.isEmpty(dto.getId())) {
+      return DeleteResponse.emptyId();
+    }
+
+    final var result = repository.deleteById(dto.getId());
+    return result ? DeleteResponse.success() : DeleteResponse.faulty();
   }
 
   @Override
